minor optimzations in screen draw functions

+++ include/screen.h
@@ -27,9 +27,10 @@
  */
 #ifndef SCREEN_H
 #define SCREEN_H
-void clearArea( int xpos, int ypos, int xsize, int ysize);
+#include "defines.h"
+void clearArea(BYTE xpos, BYTE ypos, BYTE xsize, BYTE ysize);
 void newscreen(char * title);
-void drawFrame( char * title, int xpos, int ypos, int xsize, int ysize);
+void drawFrame( char * title, BYTE xpos, BYTE ypos, BYTE xsize, BYTE ysize);
 void initScreen( unsigned char border, unsigned char bg, unsigned char text);
 void exitScreen(void);
 #endif
